+++ src/common/src/hash/consistent_hash/mapping.rs
@@ -141,7 +141,8 @@ impl<T: VnodeMappingItem> VnodeMapping<T> {
 
     /// Create a vnode mapping where all vnodes are mapped to the same single item.
     ///
-    /// The vnode count will be 1.
+    /// The length of the mapping will be 1, as if there's only one vnode in total.
+    /// This is to be consistent with [`VnodeBitmapExt::singleton`].
     pub fn new_single(item: T::Item) -> Self {
         Self::new_uniform(std::iter::once(item), 1)
     }
