+++ src/common/src/hash/mod.rs
@@ -19,6 +19,7 @@ mod key_v2;
 pub mod table_distribution;
 
 pub use consistent_hash::bitmap::*;
+pub use consistent_hash::compat::*;
 pub use consistent_hash::mapping::*;
 pub use consistent_hash::vnode::*;
 pub use dispatcher::{calc_hash_key_kind, HashKeyDispatcher};
