+++ src/frontend/src/catalog/table_catalog.rs
@@ -172,7 +172,20 @@ pub struct TableCatalog {
 
     pub cdc_table_id: Option<String>,
 
-    /// Can be unset.
+    /// Total vnode count of the table.
+    ///
+    /// Can be unset if the catalog is generated by the frontend and not yet persisted. This is
+    /// because the vnode count of each fragment (then internal tables) is determined by the
+    /// meta service. See also [`StreamMaterialize::derive_table_catalog`] and
+    /// [`TableCatalogBuilder::build`].
+    ///
+    /// On the contrary, if this comes from a [`PbTable`], the field must be `Some` no matter
+    /// whether the table is created before or after the version we introduced variable vnode
+    /// count support. This is because we've already handled backward compatibility during
+    /// conversion.
+    ///
+    /// [`StreamMaterialize::derive_table_catalog`]: crate::optimizer::plan_node::StreamMaterialize::derive_table_catalog
+    /// [`TableCatalogBuilder::build`]: crate::optimizer::plan_node::utils::TableCatalogBuilder::build
     pub vnode_count: Option<usize>,
 }
 
@@ -391,6 +404,9 @@ impl TableCatalog {
         self.version().map(|v| v.version_id)
     }
 
+    /// Get the total vnode count of the table.
+    ///
+    /// Panics if it's called on an incomplete (and not yet persisted) table catalog.
     pub fn vnode_count(&self) -> usize {
         self.vnode_count
             .expect("vnode count unset, are you calling on an incomplete table catalog?")
